Change shop module

+++ backend/modules/shop/models/Offers3.php
@@ -2,6 +2,7 @@
 
 namespace app\modules\shop\models;
 
+use common\models\shop\ShopCharacteristics;
 use common\models\shop\ShopGoods;
 use common\models\shop\ShopItemCharacteristics;
 use common\models\shop\ShopItems;
@@ -15,6 +16,7 @@ use common\models\shop\ShopPriceTypes;
 class Offers3 extends Model
 {
     private $offers_file;
+    private $chName = 'Характеристики';
 
     public function parser($offers_file)
     {
@@ -46,39 +48,92 @@ class Offers3 extends Model
 
             if ($itemVerificationCode) { // Товар с характеристикой
                 $shopItem = ShopItems::findOne(['verification_code' => $itemVerificationCode]);
-                if (!$shopItem) { // Обновление 1с-кодов
+                if (!$shopItem) {
                     $chFullName = $item->{'Наименование'};
                     $goodName = preg_quote($good->name);
                     $goodName = str_replace("/", "\/", $goodName);
 
+                    $chValue = preg_replace("/$goodName \(/", '', $chFullName);
+                    $chValue = preg_replace('/\)$/', '', $chValue);
+
+                    $ch = ShopItemCharacteristics::find()
+                        ->innerJoinWith('shopItem')
+                        ->where(['name' => $chValue])
+                        ->andWhere(['shop_goods_id' => $good->id])
+                        ->one();
+
+                    if ($ch) { // Обновление 1с-кодов
+                        $shopItem = ShopItems::findOne($ch->shop_items_id);
+                        $shopItem->verification_code = $itemVerificationCode;
+                        $shopItem->save();
+                    } else {
+                        $shopItem = new ShopItems();
+                        $shopItem->shop_goods_id = $good->id;
+                        $shopItem->verification_code = $itemVerificationCode;
+                        $shopItem->state = $item->{'ПометкаУдаления'} == 'false' ? 1 : 0;
+                        $shopItem->save();
+
+                        $shopCh = ShopCharacteristics::findOne(['name' => $this->chName]);
+                        if (!$shopCh) {
+                            $shopCh = new ShopCharacteristics();
+                            $shopCh->name = $this->chName;
+                            $shopCh->save();
+                        }
+
+                        $shopItemCh = new ShopItemCharacteristics();
+                        $shopItemCh->shop_items_id = $shopItem->id;
+                        $shopItemCh->shop_characteristics_id = $shopCh->id;
+                        $shopItemCh->name = $chValue;
+                        $shopItemCh->state = $item->{'ПометкаУдаления'} == 'false' ? 1 : 0;
+                        $shopItemCh->save();
+                    }
+                }
+
+                $priceItems = ShopPriceItem::find()->innerJoinWith('shopItem')->where(['shop_items.verification_code' => $itemVerificationCode])->all();
+                if ($priceItems) ShopPriceItem::deleteAll(['id' => ArrayHelper::getColumn($priceItems, 'id')]);
+            }
+        }
+
+    }
+
+    function parserOffersDef($offersSxe) {
+        foreach ($offersSxe as $item) {
+            $itemVerificationCode = false;
+            if (preg_match('/(.+)#(.+)/', $item->{'Ид'}, $matches)) {
+                $goodVerificationCode = strval($matches[1]);
+                $itemVerificationCode = strval($matches[2]);
+            } else {
+                $goodVerificationCode = strval($item->{'Ид'});
+            }
+            $good = ShopGoods::findOne(['verification_code' => $goodVerificationCode]);
+            if (!$good) continue;
+            if ($itemVerificationCode) { // Товар с характеристикой
+                $shopItem = ShopItems::findOne(['verification_code' => $itemVerificationCode]);
+                if (!$shopItem) { // Обновление 1с-кодов
+                    $chFullName = $item->{'Наименование'};
+                    $goodName = preg_quote($good->name);
+                    $goodName = str_replace("/", "\/", $goodName);
                     $chName = preg_replace("/$goodName \(/", '', $chFullName);
                     $chName = preg_replace('/\)/', '', $chName);
-
                     $ch = ShopItemCharacteristics::find()
                         ->innerJoinWith('shopItem')
                         ->where(['name' => $chName])
                         ->andWhere(['shop_goods_id' => $good->id])
                         ->one();
-
                     if ($ch) {
                         $shopItem = ShopItems::findOne($ch->shop_items_id);
                         $shopItem->verification_code = $itemVerificationCode;
                         $shopItem->save();
                     }
                 }
-
                 $priceItems = ShopPriceItem::find()->innerJoinWith('shopItem')->where(['shop_items.verification_code' => $itemVerificationCode])->all();
                 if ($priceItems) ShopPriceItem::deleteAll(['id' => ArrayHelper::getColumn($priceItems, 'id')]);
             } else {}
-
             $priceGood = ShopPriceGood::find()->innerJoinWith('shopGood')->where(['shop_goods.verification_code' => $goodVerificationCode])->all();
             if ($priceGood) ShopPriceGood::deleteAll(['id' => ArrayHelper::getColumn($priceGood, 'id')]);
         }
-
-
         $priceTypes = ArrayHelper::map(ShopPriceTypes::find()->all(), 'verification_code', 'id');
         $goodMinPrice = array();
-
         foreach ($offersSxe as $itemSxe) {
             if (preg_match('/(.+)#(.+)/', $itemSxe->{'Ид'}, $matches)) {
                 $goodVrfCode = strval($matches[1]);
@@ -86,9 +141,7 @@ class Offers3 extends Model
             } else {
                 $goodVrfCode = strval($itemSxe->{'Ид'});
             }
-
             if (!isset($goodMinPrice[$goodVrfCode])) $goodMinPrice[$goodVrfCode] = array();
-
             foreach ($itemSxe->{'Цены'}->{'Цена'} as $itemPrice) {
                 $priceVrfCode = strval($itemPrice->{'ИдТипаЦены'});
                 if (!isset($goodMinPrice[$goodVrfCode][$priceTypes[$priceVrfCode]])) {
@@ -96,7 +149,6 @@ class Offers3 extends Model
                 } else if ($goodMinPrice[$goodVrfCode][$priceTypes[$priceVrfCode]] > floatval($itemPrice->{'ЦенаЗаЕдиницу'})) {
                     $goodMinPrice[$goodVrfCode][$priceTypes[$priceVrfCode]] = floatval($itemPrice->{'ЦенаЗаЕдиницу'});
                 }
-
                 if ($itemVrfCode) {
                     $itemShop = ShopItems::findOne(['verification_code' => $itemVrfCode]);
                     if ($itemShop) {
@@ -109,7 +161,6 @@ class Offers3 extends Model
                 }
             }
         }
-
         foreach ($goodMinPrice as $good_verification_code => $goodPrices) {
             $good = ShopGoods::findOne(['verification_code' => $good_verification_code]);
             if ($good) {
